# resources/skg-if-openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0-alpha
  title: SKG-IF OpenAPI
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at : https://skg-if.github.io/api/

    # Note for implementers :
      - Please refer to [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewer :
        - StopLight/Elements Viewer [Current/last SKG-IF OpenAPI version](https://elements-demo.stoplight.io/?spec=https://w3id.org/skg-if/api/skg-if-openapi.yaml).
        - Swagger : [Current/last SKG-IF OpenAPI version](https://editor-next.swagger.io/)
...
servers:
  - url: '{protocol}://{server}:{port}/{skg_if_api_path}'
    variables:
      protocol:
        enum:
          - "https"
          - "http"
        default: "https"
        description: service provider protocol.
      server:
        default: example.com
...
      port:
        default: "8443"
...
paths:
...
  '/persons/{short_local_identifier}':
    get:
      tags:
          - Person
      summary: Get person by id
      description: |
          Get `person` by id. See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:person ) .\
      operationId: getPersonById
      parameters:
        - $ref : '#/components/parameters/shortLocalIdPathParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Person"
              examples:
                PersonEx01:
                  summary: Josiah Carberry person
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "http://example.com/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/persons/pers-c66c6-38be-4d5f-85db-d44c9f869333
                        entity_type: "person"
                        identifiers:
                          - scheme: orcid
                            value: 0000-0002-1825-0097
                        given_name: "Josiah"
                        family_name: "Carberry"
                        name: "Josiah Carberry"
                PersonEx02:
                  summary: John Doe person on-the-fly id
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "http://example.com/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/persons/otf_126848135_prod-1-pers-1
                        entity_type: "person"
                        name: "John Doe"
        '404':
          description: |
            Error if entity does not exists
...
  '/grants/{short_local_identifier}':
    get:
      tags:
          - Grant
      summary: Get grant by id
      description: |
          Get single `grant`. See definition in SKG-IF [Grant](https://skg-if.github.io/interoperability-framework/docs/grant.html) ( entity_type:grant ).
      operationId: getGrantById
      parameters:
        - $ref : '#/components/parameters/shortLocalIdPathParam'
      responses:
        '200':
          description: Success
          content:
            # not 'application/json-ld' to be compatible with StopLight PRISM tool
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Grant"
              examples:
                GrantEx01:
                  summary: GraspOS grant
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/grants/grant-c66c6-38be-4d5f-85db-d44c9f869333
                        entity_type: "grant"
                        identifiers:
                          - scheme: "doi"
                            value: "https://doi.org/10.3030/101095129"
                        grant_number: "101095129"
                        titles:
                          "en": "GraspOS: next Generation Research Assessment to Promote Open Science"
                        abstracts: {
                            "en": "GraspOS aims to build and operate a data infrastructure to support the policy reforms..."
                        }
                        acronym: "GraspOS"
                        funding_agency:
                          local_identifier: http://example.com/skg-if/api/grants/otf___1730027051396___org-1
                          entity_type: "organisation"
                          identifiers:
                            - scheme: "ror"
                              value: "https://ror.org/05gq02987"
                          name: "Brown University"
                          short_name: "BU"
                          country: "US"
                          website: "https://www.brown.edu/"
                        funding_stream: "Horizon Europe"
                        currency: "EUR"
                        funded_amount: 50000
                        duration:
                          start: "2023-01-01T00:00:00Z"
                          end: "2025-12-31T23:59:59Z"
                        website: "https://graspos.eu"
                        beneficiaries:
                          - local_identifier: http://example.com/skg-if/api/organisations/otf___1730027051396___org-2
                            entity_type: "organisation"
                            identifiers:
                              - scheme: "ror"
                                value: "https://ror.org/05gq02987"
                            name: "Brown University"
                            short_name: "BU"
                            country: "US"
                            website: "https://www.brown.edu/"
                        contributions:
                          - by:
                              local_identifier: http://example.com/skg-if/api/persons/otf___1730027051396___pers-1
                              entity_type: "person"
                              identifiers:
                                - scheme: "orcid"
                                  value: "0000-0002-1825-0097"
                              given_name: "Josiah"
                              family_name: "Carberry"
                              name: "Josiah Carberro"
                              declared_affiliations:
                                - local_identifier": "otf___1730027051396___org-2"
                                  entity_type": "organisation"
                                  identifiers:
                                    - scheme: "ror"
                                      value: "https://ror.org/05gq02987"
                                  name: "Brown University"
                                  short_name: "BU"
                                  country: "US"
                                  website: "https://www.brown.edu/"
                              roles:
                                - "co-applicant"
        '404':
          description: |
            Error if entity does not exists
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    shortLocalIdPathParam:
      name: short_local_identifier
      in: path
      description: |
        entity id. URL suffix used in OpenAPI get by id operations. \
        `https://[your-server-root-skg-if-url]/[products|persons|organisations|datasources|venues|topics]/{short_local_identifier}` \
        example `http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9`
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
...
  schemas:
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/my-skg-acronym" }`'
      properties:
        "@base":
          type: string
          description : |
            Default base URL specific to the implementer.\
            common format : `https://w3id.org/skg-if/sandbox/{my-skg-acronym}` \
            example : `https://w3id.org/skg-if/sandbox/openaire/`
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://w3id.org/skg-if/sandbox/openaire/"
            - "https://w3id.org/skg-if/sandbox/cessda/"
            - "https://w3id.org/skg-if/sandbox/foobar/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg-if/{{random.alpha(3)}}']